+++ structurizr-core/src/com/structurizr/model/Container.java
@@ -174,7 +174,7 @@ public final class Container extends StaticStructureElement {
      */
     @Override
     public String getCanonicalName() {
-        return getParent().getCanonicalName() + CANONICAL_NAME_SEPARATOR + formatForCanonicalName(getName());
+        return new CanonicalNameGenerator().generate(this);
     }
 
     @Override