+++ structurizr-core/src/com/structurizr/model/InfrastructureNode.java
@@ -60,21 +60,6 @@ public final class InfrastructureNode extends DeploymentElement {
         return getModel().addRelationship(this, destination, description, technology, interactionStyle, tags);
     }
 
-    /**
-     * Gets the parent deployment node.
-     *
-     * @return  the parent DeploymentNode, or null if there is no parent
-     */
-    @Override
-    @JsonIgnore
-    public Element getParent() {
-        return parent;
-    }
-
-    void setParent(DeploymentNode parent) {
-        this.parent = parent;
-    }
-
     public String getTechnology() {
         return technology;
     }
@@ -90,7 +75,7 @@ public final class InfrastructureNode extends DeploymentElement {
 
     @Override
     public String getCanonicalName() {
-        return getParent().getCanonicalName() + CANONICAL_NAME_SEPARATOR + formatForCanonicalName(getName());
+        return new CanonicalNameGenerator().generate(this);
     }
 
 }