+++ structurizr-core/src/com/structurizr/model/SoftwareSystem.java
@@ -139,7 +139,7 @@ public final class SoftwareSystem extends StaticStructureElement {
      */
     @Override
     public String getCanonicalName() {
-        return CANONICAL_NAME_SEPARATOR + formatForCanonicalName(getName());
+        return new CanonicalNameGenerator().generate(this);
     }
 
     @Override